requirements.txt
django
dataclasses
requests
gunicorn
pymongo
dnspython
django-cors-headers
pyjwt
fbdown
TikTokApi